feat(sanitize): refactor
